Move `sign`/`decrypt`/`import`/`attest` to the `key` module
These are crypto key-related functions and are better factored under
this module.
